+++ lib/internal/modules/esm/translators.js
@@ -8,7 +8,8 @@ const {
 } = require('internal/modules/cjs/helpers');
 const CJSModule = require('internal/modules/cjs/loader');
 const internalURLModule = require('internal/url');
-const createDynamicModule = require('internal/modules/esm/CreateDynamicModule');
+const createDynamicModule = require(
+  'internal/modules/esm/create_dynamic_module');
 const fs = require('fs');
 const { _makeLong } = require('path');
 const { SafeMap } = require('internal/safe_globals');
