+++ test/es-module/test-esm-loader-modulemap.js
@@ -7,10 +7,11 @@
 const common = require('../common');
 
 const { URL } = require('url');
-const Loader = require('internal/modules/esm/Loader');
-const ModuleMap = require('internal/modules/esm/ModuleMap');
-const ModuleJob = require('internal/modules/esm/ModuleJob');
-const createDynamicModule = require('internal/modules/esm/CreateDynamicModule');
+const Loader = require('internal/modules/esm/loader');
+const ModuleMap = require('internal/modules/esm/module_map');
+const ModuleJob = require('internal/modules/esm/module_job');
+const createDynamicModule = require(
+  'internal/modules/esm/create_dynamic_module');
 
 const stubModuleUrl = new URL('file://tmp/test');
 const stubModule = createDynamicModule(['default'], stubModuleUrl);
