+++ test/message/esm_display_syntax_error.out
@@ -3,4 +3,4 @@ file:///*/test/message/esm_display_syntax_error.mjs:3
 await async () => 0;
 ^^^^^
 SyntaxError: Unexpected reserved word
-    at translators.set (internal/modules/esm/Translators.js:*:*)
+    at translators.set (internal/modules/esm/translators.js:*:*)
